Clean up imports in chapter 5 and 6
Wildcard imports were removed. Imports through the whole chapter 6
were made consistent. Clean up was made taking in account code
readability and reusability.
core and protocols are incredibly used, hence are directly imported.
This permits to (depending on the notebook):
- Avoid writing pyrosetta.rosetta before every statement
(it is hard to limit a line to 78 characters when a function call
is 70 characters long. This removes 17 while enhancing readability.)
- Avoid extensive imports by directly calling functions used once or
twice, allowing more flexibility for the programmer, while the
location of a function is clearly given for the reader, making it
easier to reuse the line (instead of having to go check the imports,
or worst try every modules that were wildcard imported).

If several functions of a module where used in a notebook, then it
was imported (ex: protocols.antibody and loops in notebook 06.02).

Importing task from core.pack permits to clearly show objects as
TaskOperator (task.operator.Xxxx) while being able to call
task.TaskFactory.

Importing core.select.residue_selector as selectors is a practice
initially found in notebook 06.02: it was merely expended to others,
again to limit the number of char per line and for consistency sake.

Several loose PEP8 adjustments were manually made.

Cells were tested in a copy of each notebook, to avoid recording
metadata or output changes.