Remove Second Declaration of FastRelax
Creating the second fast relax object without setting its scorefunction and attempting to apply it will create a segmentation fault. Also the import for fast relax is above in the same cell so there is no need to use the whole rosetta.protocols.relax.FastRelax() syntax to create a new fast relax object.